@@ -37,13 +36,13 @@
         "dotenv": "10.0.0",
         "dotenv-expand": "5.1.0",
         "fs-extra": "^10.0.0",
-        "http-proxy-middleware": "^2.0.0",
+        "http-proxy-middleware": "^2.0.1",
         "identity-obj-proxy": "3.0.0",
-        "jest": "27.0.4",
-        "jest-circus": "27.0.4",
-        "jest-resolve": "27.0.4",
+        "jest": "27.1.1",
+        "jest-circus": "27.1.1",
+        "jest-resolve": "27.1.1",
         "jest-watch-typeahead": "0.6.4",
-        "next": "11.0.0",
+        "next": "11.1.2",
         "next-transpile-modules": "^8.0.0",
         "nextjs-redirect": "^5.0.2",
         "nprogress": "^0.2.0",
@@ -56,29 +55,29 @@
         "resolve": "1.20.0",
         "resolve-url-loader": "^4.0.0",
         "semver": "7.3.5",
-        "styled-components": "^5.3.0",
-        "ts-loader": "^9.2.3",
+        "styled-components": "^5.3.1",
+        "ts-loader": "^9.2.5",
         "ts-pnp": "1.2.0",
-        "typescript": "^4.3.4",
+        "typescript": "^4.4.3",
         "use-interval": "^1.3.0",
-        "web-vitals": "^2.0.1",
-        "webpack": "^5.39.1"
+        "web-vitals": "^2.1.0",
+        "webpack": "^5.52.1"
     },
     "devDependencies": {
-        "@types/jest": "^26.0.23",
+        "@types/jest": "^27.0.1",
         "@types/nprogress": "^0.2.0",
-        "@types/react": "^17.0.11",
-        "@types/react-beautiful-dnd": "^13.0.0",
-        "@types/react-dom": "^17.0.8",
-        "@types/react-router": "^5.1.15",
-        "@types/react-router-dom": "^5.1.7",
-        "@types/styled-components": "^5.1.10",
-        "eslint-config-next": "^11.0.0",
+        "@types/react": "^17.0.20",
+        "@types/react-beautiful-dnd": "^13.1.2",
+        "@types/react-dom": "^17.0.9",
+        "@types/react-router": "^5.1.16",
+        "@types/react-router-dom": "^5.1.8",
+        "@types/styled-components": "^5.1.14",
+        "eslint-config-next": "^11.1.2",
         "eslint-plugin-jsx-a11y": "^6.4.1",
-        "eslint-plugin-react": "^7.24.0",
+        "eslint-plugin-react": "^7.25.1",
         "eslint-plugin-react-hooks": "^4.2.0",
-        "eslint-plugin-testing-library": "^4.6.0",
-        "eslint-webpack-plugin": "^2.5.4"
+        "eslint-plugin-testing-library": "^4.12.2",
+        "eslint-webpack-plugin": "^3.0.1"
     },
     "scripts": {
         "circular": "madge --circular --ts-config ./tsconfig.json --extensions ts ./",
