@@ -24,52 +24,52 @@
         "testsuite": "ultra -r yarn testsuite",
         "up": "ultra -r --root ncu -u --packageFile package.json -x @types/node",
         "pub": "yarn lerna publish --no-private",
-        "client": "cd packages/client && yarn client",
-        "libs-fix": "mkdir -p libs/node_modules/.bin && cd libs/node_modules/.bin && ln -s ../../../node_modules/.bin/* ."
+        "client": "cd packages/client && yarn client"
     },
     "resolutions": {
-        "typescript": "^4.3.4"
+        "typescript": "^4.4.3"
     },
     "devDependencies": {
-        "@babel/node": "^7.14.5",
+        "@babel/node": "^7.15.4",
         "@babel/plugin-proposal-class-properties": "^7.14.5",
-        "@babel/plugin-proposal-decorators": "^7.14.5",
-        "@babel/plugin-proposal-object-rest-spread": "^7.14.5",
-        "@babel/preset-typescript": "^7.14.5",
-        "@effect-ts/jest": "^0.28.4",
-        "@effect-ts/tracing-plugin": "^0.11.0",
-        "@types/jest": "^26.0.23",
-        "@types/node": "^15.6.1",
-        "@typescript-eslint/eslint-plugin": "^4.27.0",
-        "@typescript-eslint/parser": "^4.27.0",
-        "babel-jest": "^27.0.2",
+        "@babel/plugin-proposal-decorators": "^7.15.4",
+        "@babel/plugin-proposal-object-rest-spread": "^7.15.6",
+        "@babel/preset-typescript": "^7.15.0",
+        "@effect-ts/jest": "^0.29.0",
+        "@effect-ts/tracing-plugin": "^0.14.8",
+        "@types/jest": "^27.0.1",
+        "@types/lodash": "^4.14.172",
+        "@types/node": "^16.9.1",
+        "@typescript-eslint/eslint-plugin": "^4.31.0",
+        "@typescript-eslint/parser": "^4.31.0",
+        "babel-jest": "^27.1.1",
         "babel-plugin-module-resolver": "^4.1.0",
-        "concurrently": "^6.2.0",
-        "core-js": "^3.14.0",
-        "esbuild": "^0.12.9",
+        "concurrently": "^6.2.1",
+        "core-js": "^3.17.3",
+        "esbuild": "^0.12.26",
         "esbuild-node": "^0.4.0",
-        "esbuild-runner": "^2.2.0",
-        "eslint": "^7.28.0",
+        "esbuild-runner": "^2.2.1",
+        "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-import-resolver-typescript": "^2.4.0",
         "eslint-import-resolver-webpack": "^0.13.1",
-        "eslint-plugin-import": "^2.23.4",
-        "eslint-plugin-prettier": "^3.4.0",
+        "eslint-plugin-import": "^2.24.2",
+        "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-simple-import-sort": "^7.0.0",
         "eslint-plugin-sort-destructure-keys": "^1.3.5",
-        "fast-check": "^2.16.0",
-        "jest": "^27.0.4",
+        "fast-check": "^2.17.0",
+        "jest": "^27.1.1",
         "lerna": "^4.0.0",
-        "madge": "^4.0.2",
-        "nodemon": "^2.0.7",
-        "npm-check-updates": "^11.7.1",
+        "madge": "^5.0.1",
+        "nodemon": "^2.0.12",
+        "npm-check-updates": "^11.8.5",
         "patch-package": "^6.4.7",
         "postinstall-postinstall": "^2.1.0",
-        "prettier": "^2.3.1",
-        "ts-node": "^10.0.0",
-        "tsconfig-paths": "^3.9.0",
+        "prettier": "^2.4.0",
+        "ts-node": "^10.2.1",
+        "tsconfig-paths": "^3.11.0",
         "ttypescript": "^1.5.12",
-        "typescript": "^4.3.4",
+        "typescript": "^4.4.3",
         "ultra-runner": "^3.10.5"
     }
 }
