@@ -1,12 +1,12 @@
 /* eslint-disable react/display-name */
 import * as MO from "@effect-ts-app/core/Schema"
 import * as O from "@effect-ts/core/Option"
-import { Box, List, ListItem, ListItemIcon, ListItemText } from "@material-ui/core"
-import Balance from "@material-ui/icons/AccountBalance"
-import CalendarToday from "@material-ui/icons/CalendarToday"
-import Calendar from "@material-ui/icons/CalendarViewMonth"
-import Home from "@material-ui/icons/Home"
-import Star from "@material-ui/icons/Star"
+import { Box, List, ListItem, ListItemIcon, ListItemText } from "@mui/material"
+import Balance from "@mui/icons-material/AccountBalance"
+import CalendarToday from "@mui/icons-material/CalendarToday"
+import Calendar from "@mui/icons-material/CalendarViewMonth"
+import Home from "@mui/icons-material/Home"
+import Star from "@mui/icons-material/Star"
 import Link from "next/link"
 import React from "react"
 import { Draggable, Droppable } from "react-beautiful-dnd"
