@@ -1,5 +1,5 @@
-import { Box, Checkbox, IconButton } from "@material-ui/core"
-import Remove from "@material-ui/icons/Remove"
+import { Box, Checkbox, IconButton } from "@mui/material"
+import Remove from "@mui/icons-material/Remove"
 import React from "react"
 import { Draggable } from "react-beautiful-dnd"
 
