@@ -1,9 +1,9 @@
 import * as A from "@effect-ts-app/core/Array"
 import * as O from "@effect-ts/core/Option"
-import { Box, Card, Checkbox } from "@material-ui/core"
-import Alarm from "@material-ui/icons/Alarm"
-import CalendarToday from "@material-ui/icons/CalendarToday"
-import Today from "@material-ui/icons/Today"
+import { Box, Card, Checkbox } from "@mui/material"
+import Alarm from "@mui/icons-material/Alarm"
+import CalendarToday from "@mui/icons-material/CalendarToday"
+import Today from "@mui/icons-material/Today"
 import { datumEither } from "@nll/datum"
 import React from "react"
 import { Draggable } from "react-beautiful-dnd"
