@@ -6,10 +6,10 @@ import { flow } from "@effect-ts/core/Function"
 import * as O from "@effect-ts/core/Option"
 import * as ORD from "@effect-ts/core/Ord"
 import { Ord } from "@effect-ts/core/Ord"
-import { Box, Button, IconButton, Typography } from "@material-ui/core"
-import ArrowDown from "@material-ui/icons/ArrowDownward"
-import ArrowUp from "@material-ui/icons/ArrowUpward"
-import Refresh from "@material-ui/icons/Refresh"
+import { Box, Button, IconButton, Typography } from "@mui/material"
+import ArrowDown from "@mui/icons-material/ArrowDownward"
+import ArrowUp from "@mui/icons-material/ArrowUpward"
+import Refresh from "@mui/icons-material/Refresh"
 import { datumEither } from "@nll/datum"
 import React from "react"
 import useInterval from "use-interval"
