@@ -1,9 +1,9 @@
 import { UserProvider } from "@auth0/nextjs-auth0"
-import { StyledEngineProvider, useMediaQuery } from "@material-ui/core"
-import { createTheme, ThemeProvider } from "@material-ui/core/styles"
-import AdapterDateFns from "@material-ui/lab/AdapterDateFns"
-import LocalizationProvider from "@material-ui/lab/LocalizationProvider"
-import { StylesProvider } from "@material-ui/styles"
+import AdapterDateFns from "@mui/lab/AdapterDateFns"
+import LocalizationProvider from "@mui/lab/LocalizationProvider"
+import { StyledEngineProvider, useMediaQuery } from "@mui/material"
+import { createTheme, ThemeProvider } from "@mui/material/styles"
+import { StylesProvider } from "@mui/styles"
 import { useRouter } from "next/router"
 import NProgress from "nprogress"
 import React, { useEffect, useMemo } from "react"
