@@ -2,9 +2,10 @@
 import * as T from "@effect-ts/core/Effect"
 import * as Test from "@effect-ts/jest/Test"
 import * as O from "@effect-ts-app/core/Option"
+import { ReasonableString } from "@effect-ts-app/core/Schema"
 import { TaskAudits, TaskEvents } from "@effect-ts-demo/todo-types/Task"
 
-import { reasonableStringUnsafe, testUser } from "@/test.helpers"
+import { testUser } from "@/test.helpers"
 
 import { createTask_ } from "./Create"
 
@@ -18,7 +19,7 @@ it("returns domain events", () =>
       listId: "inbox",
       myDay: O.none,
       isFavorite: false,
-      title: reasonableStringUnsafe("Test task"),
+      title: ReasonableString.unsafe("Test task"),
     })
 
     expect(events).toEqual([
@@ -32,7 +33,7 @@ it("adds an Audit on creation", () =>
       listId: "inbox",
       myDay: O.none,
       isFavorite: false,
-      title: reasonableStringUnsafe("Test task"),
+      title: ReasonableString.unsafe("Test task"),
     })
 
     expect(task.auditLog[0]).toEqual(
