@@ -1,18 +1,12 @@
-import {
-  emailUnsafe,
-  phoneNumberUnsafe,
-  reasonableStringUnsafe,
-  userIdUnsafe,
-} from "@effect-ts-demo/todo-client/test.helpers"
+import { Email, PhoneNumber, ReasonableString } from "@effect-ts-app/core/Schema"
+import { UserId } from "@effect-ts-demo/todo-types/"
 import { User } from "@effect-ts-demo/todo-types/User"
-
-export * from "@effect-ts-demo/todo-client/test.helpers"
 
 export function testUser() {
   return new User({
-    id: userIdUnsafe("2"),
-    name: reasonableStringUnsafe("Patroklos"),
-    email: emailUnsafe("some@test.com"),
-    phoneNumber: phoneNumberUnsafe("555-444-123"),
+    id: UserId.unsafe("2"),
+    name: ReasonableString.unsafe("Patroklos"),
+    email: Email.unsafe("some@test.com"),
+    phoneNumber: PhoneNumber.unsafe("555-444-123"),
   })
 }
