@@ -2,9 +2,9 @@ import { Post, prop, useClassNameForSchema } from "@effect-ts-app/core/Schema"
 import { TaskListId, UserId } from "@effect-ts-demo/todo-types"
 
 @useClassNameForSchema
-export default class AddTaskListMember extends Post(
+export class AddTaskListMemberRequest extends Post(
   "/lists/:id/members"
-)<AddTaskListMember>()({
+)<AddTaskListMemberRequest>()({
   id: prop(TaskListId),
   memberId: prop(UserId),
 }) {}
