@@ -2,6 +2,6 @@ import { Delete, prop, useClassNameForSchema } from "@effect-ts-app/core/Schema"
 import { TaskListId } from "@effect-ts-demo/todo-types"
 
 @useClassNameForSchema
-export default class RemoveTaskList extends Delete("/lists/:id")<RemoveTaskList>()({
+export class RemoveTaskListRequest extends Delete("/lists/:id")<RemoveTaskListRequest>()({
   id: prop(TaskListId),
 }) {}
