@@ -1,7 +1,7 @@
 import {
   date,
   Model,
-  nullable,
+  optionFromNull,
   Post,
   prop,
   useClassNameForSchema,
@@ -9,13 +9,13 @@ import {
 import { Task, TaskId } from "@effect-ts-demo/todo-types"
 
 @useClassNameForSchema
-export default class CreateTask extends Post("/tasks")<CreateTask>()({
+export class CreateTaskRequest extends Post("/tasks")<CreateTaskRequest>()({
   ...Task.include(({ isFavorite, listId, title }) => ({
     listId,
     title,
     isFavorite,
   })),
-  myDay: prop(nullable(date)),
+  myDay: prop(optionFromNull(date)),
 }) {}
 
 export class Response extends Model<Response>()({ id: prop(TaskId) }) {}
