@@ -1,15 +1,15 @@
 import * as O from "@effect-ts-app/core/Option"
-import { reasonableStringUnsafe } from "@effect-ts-app/core/test.helpers"
+import { ReasonableString } from "@effect-ts-app/core/Schema"
+import { UserId } from "@effect-ts-demo/todo-types"
 
-import { userIdUnsafe } from "../test.helpers"
 import { TaskCreated } from "./audit"
 import { Task } from "./Task"
 
 describe("Constructor", () => {
   describe("with task", () => {
     const t = new Task({
-      createdBy: userIdUnsafe("1"),
-      title: reasonableStringUnsafe("yay"),
+      createdBy: UserId.unsafe("1"),
+      title: ReasonableString.unsafe("yay"),
       completed: O.some(new Date(2040, 12, 24)),
     })
 
@@ -31,8 +31,8 @@ describe("Constructor", () => {
       expect(
         () =>
           new Task({
-            createdBy: userIdUnsafe("1"),
-            title: reasonableStringUnsafe("yay"),
+            createdBy: UserId.unsafe("1"),
+            title: ReasonableString.unsafe("yay"),
             completed: O.some(new Date(1983, 12, 24)),
           })
       ).toThrow()
