@@ -1,19 +1,19 @@
 import {
-  longString,
+  LongString,
   Model,
   ParsedShapeOf,
   prop,
-  reasonableString,
+  ReasonableString,
 } from "@effect-ts-app/core/Schema"
 
-export const FileName = reasonableString
+export const FileName = ReasonableString
 export type FileName = ParsedShapeOf<typeof FileName>
 
 // TODO
-export const Url = longString
+export const Url = LongString
 export type Url = ParsedShapeOf<typeof Url>
 
-export const MimeType = reasonableString
+export const MimeType = ReasonableString
 export type MimeType = ParsedShapeOf<typeof MimeType>
 
 export class Attachment extends Model<Attachment>()({
